изменение меню, страниц, взаимодействия меню и страниц
