--- docs/docs/architecture.md
@@ -1,22 +1,0 @@
-# Архитектура
-
-GeoOffice построен по модульному принципу, ориентирован на расширяемость и переиспользование кода.
-
-- **app.py** — точка входа, инициализация приложения, страниц, сервисов и компонентов
-- **pages/** — отдельные страницы (наследуют BasePage), реализуют UI и взаимодействие с сервисами
-- **components/** — переиспользуемые UI-компоненты (меню, поиск, хлебные крошки, статистика)
-- **services/** — бизнес-логика (работа с картограммой, масштабами, координатами и др.)
-- **models/** — структуры данных (CartogramData, ScaleCalculation, Coordinate и др.)
-- **utils/** — утилиты (FileUtils, LoggerConfig, ScaleUtils, CoordinateUtils)
-
-## Взаимодействие между слоями
-- Страницы используют сервисы для бизнес-логики и модели для хранения данных
-- Компоненты получают данные через страницы или напрямую из сервисов
-- Все слои используют утилиты для работы с файлами, логированием и расчетами
-
-## Зависимости
-- Нет жёсткой связи между страницами и компонентами — через DI (передача app)
-- Сервисы не зависят от UI, только от моделей и утилит
-- Модели не зависят ни от чего, кроме стандартных библиотек и dataclasses
-
-Архитектура позволяет легко добавлять новые страницы, сервисы и компоненты без изменения существующего кода.
