--- docs/docs/autocad.md
@@ -1,18 +1,0 @@
-# AutoCAD
-
-Раздел для работы с интеграцией и плагинами AutoCAD. Позволяет автоматизировать оформление чертежей и экспорт данных из GeoOffice.
-
-## Основные функции
-- Установка плагина для мультивыносок (автоматизация подписей)
-- Установка плагина для подписей координат
-- Экспорт данных из GeoOffice в DXF/DWG для AutoCAD
-- Автоматизация работы с чертежами (генерация, обновление, пакетная обработка)
-
-## Используемые сервисы и утилиты
-- `AutoCADService` — взаимодействие с AutoCAD, генерация DXF
-- `FileUtils` — подготовка файлов для экспорта
-
-## Пример сценария
-1. Сгенерировать данные в GeoOffice (координаты, картограммы)
-2. Экспортировать в DXF/DWG
-3. Открыть файл в AutoCAD и применить плагины для оформления
