--- docs/docs/cartogram.md
@@ -1,27 +1,0 @@
-# Картограмма
-
-Раздел для создания, редактирования и экспорта картограмм. Позволяет автоматизировать подготовку картографических материалов для отчетов и кадастровых работ.
-
-## Основные возможности
-- Загрузка DXF-файлов с полигонами
-- Автоматическое определение системы координат
-- Создание сетки по заданному размеру ячейки
-- Генерация и экспорт списка номенклатур
-- Сохранение сетки в DXF
-
-## Используемые сервисы и модели
-- `CartogramService` — парсинг, генерация сетки, расчет номенклатур, экспорт
-- `CartogramData` — модель данных картограммы
-
-## Пример использования
-```python
-from src.services.cartogram_service import CartogramService
-mp = CartogramService.parse_dxf('input.dxf')
-grid = CartogramService.create_grid(mp, 250)
-CartogramService.export_grid_to_dxf(grid, 'output.dxf')
-```
-
-## Сценарии применения
-- Подготовка картограмм для отчетов и кадастровых работ
-- Автоматизация экспорта в DXF для передачи в AutoCAD
-- Генерация сетки и расчет номенклатур для анализа участков
