--- docs/docs/components_breadcrumbs.md
@@ -1,23 +1,0 @@
-# Компонент: Хлебные крошки
-
-Компонент для отображения навигации по страницам (`Breadcrumbs`). Помогает пользователю ориентироваться в структуре приложения.
-
-## Основные возможности
-- Отображение текущей категории и страницы
-- Быстрый переход к разделам
-- Интерактивные элементы (кнопки, ссылки)
-- Индикация активного раздела
-
-## Структура и методы
-- `Breadcrumbs(app)` — инициализация компонента
-- `update_breadcrumbs(path: List[str])` — обновление навигационной цепочки
-
-## Пример использования
-```python
-breadcrumbs = Breadcrumbs(app)
-breadcrumbs.update_breadcrumbs(['Документы', 'Картограмма'])
-```
-
-## Сценарии применения
-- Навигация по вложенным разделам
-- Быстрый возврат к предыдущим страницам
