--- docs/docs/components_menu.md
@@ -1,27 +1,0 @@
-# Компонент: Меню
-
-Меню с категориями и подпрограммами, реализовано в `CategorizedMenu`. Используется для навигации по основным разделам приложения.
-
-## Основные возможности
-- Категории: Главная, Документы, Геодезические работы, Специализированные инструменты, Система
-- Поддержка поиска и статистики использования
-- Свернутое и развернутое отображение
-- Быстрый переход к функциям
-- Индикация активного раздела
-
-## Структура и методы
-- `CategorizedMenu(app)` — инициализация компонента
-- `categories` — структура категорий и пунктов меню
-- `update_menu()` — обновление состояния меню
-- Взаимодействует с компонентами поиска (`MenuSearch`) и статистики (`UsageStats`)
-
-## Пример использования
-```python
-menu = CategorizedMenu(app)
-menu.update_menu()
-```
-
-## Сценарии применения
-- Основная навигация по разделам приложения
-- Быстрый доступ к часто используемым функциям
-- Встраивание в основной и свернутый интерфейс
