--- docs/docs/components_search.md
@@ -1,26 +1,0 @@
-# Компонент: Поиск по меню
-
-Компонент для поиска функций и страниц (`MenuSearch`). Позволяет быстро находить нужные разделы и переходить к ним.
-
-## Основные возможности
-- Быстрый поиск по названию, категории и описанию функции
-- Переход к найденной странице
-- Интерактивный виджет поиска (автодополнение)
-- Интеграция с меню и статистикой
-
-## Структура и методы
-- `MenuSearch(app)` — инициализация компонента
-- `all_functions` — список всех функций для поиска
-- `search(query: str)` — поиск по функциям
-- `show_results()` — отображение результатов поиска
-
-## Пример использования
-```python
-search = MenuSearch(app)
-search.search('координаты')
-search.show_results()
-```
-
-## Сценарии применения
-- Быстрый переход к нужной функции без ручного поиска в меню
-- Интеграция в основной интерфейс и всплывающие окна
