--- docs/docs/components_stats.md
@@ -1,25 +1,0 @@
-# Компонент: Статистика использования
-
-Компонент для отображения статистики использования функций (`UsageStats`). Позволяет анализировать активность пользователей и популярность разделов.
-
-## Основные возможности
-- Подсчёт количества переходов по страницам
-- Отображение наиболее используемых функций
-- Встраивание в меню и другие компоненты
-- Сброс и экспорт статистики
-
-## Структура и методы
-- `UsageStats(app)` — инициализация компонента
-- `update_stats(page_name: str)` — обновление статистики при переходе
-- `get_most_used()` — получение наиболее популярных функций
-
-## Пример использования
-```python
-stats = UsageStats(app)
-stats.update_stats('documents')
-print(stats.get_most_used())
-```
-
-## Сценарии применения
-- Анализ востребованных функций для улучшения UX
-- Ведение внутренней аналитики приложения
