--- docs/docs/conversion.md
@@ -1,24 +1,0 @@
-# Конвертация
-
-Раздел для конвертации файлов между различными форматами (JSON, SHP, DXF/DWG) для интеграции с ГИС, CAD и другими системами.
-
-## Основные функции
-- Конвертация JSON (ГЗУ) в DXF/DWG
-- Конвертация SHP (Shape) в DXF/DWG
-- Выбор и загрузка файлов для конвертации
-- Проверка корректности формата и структуры данных
-
-## Используемые сервисы и утилиты
-- `ConversionService` — преобразование форматов, валидация данных
-- `FileUtils` — загрузка и сохранение файлов
-
-## Пример использования
-```python
-from src.services.conversion_service import ConversionService
-ConversionService.json_to_dxf('input.json', 'output.dxf')
-```
-
-## Сценарии применения
-- Подготовка данных для передачи в AutoCAD
-- Конвертация результатов геодезических работ для государственных систем
-- Автоматизация пакетной обработки файлов
