--- docs/docs/coordinates.md
@@ -1,32 +1,0 @@
-# Координаты
-
-Раздел для работы с координатами и системами координат. Позволяет вводить, форматировать, преобразовывать и экспортировать координаты для различных задач.
-
-## Основные функции
-- Ввод и форматирование координат (поддержка разных разделителей)
-- Создание файлов пересчёта координат
-- Генерация файлов для Metashape и DJI Terra
-- Проверка и преобразование форматов (например, из текстового в табличный)
-
-## Используемые сервисы и утилиты
-- `CoordinateService` — парсинг, форматирование, преобразование координат
-- `FileUtils` — сохранение и загрузка файлов
-
-## Пример использования
-```python
-from src.services.coordinate_service import CoordinateService
-coords = CoordinateService.parse('100.123 200.456\n101.789 201.654')
-formatted = CoordinateService.format(coords, sep='\t')
-print(formatted)
-```
-
-## Пример формата ввода
-```
-100.123 200.456
-101.789 201.654
-```
-
-## Сценарии применения
-- Подготовка координат для загрузки в ГИС или CAD
-- Форматирование для передачи подрядчикам
-- Автоматизация создания файлов для фотограмметрии (Metashape, DJI Terra)
