--- docs/docs/documents.md
@@ -1,23 +1,0 @@
-# Документы
-
-Раздел предназначен для управления геодезической документацией, создания отчетов, картограмм и работы с шаблонами документов.
-
-## Основные функции
-- Создание отчетов о геодезических работах (на основе шаблонов)
-- Создание и экспорт картограмм
-- Импорт и экспорт файлов (JSON, DXF, SHP)
-- Использование и редактирование шаблонов документов
-- Архивирование и восстановление документов
-
-## Используемые сервисы и утилиты
-- `FileUtils` — загрузка, сохранение, проверка файлов
-- `DocumentService` — генерация и обработка документов
-
-## Пример сценария
-1. Импортировать исходные данные (JSON, DXF)
-2. Сгенерировать отчет или картограмму по шаблону
-3. Экспортировать результат в нужном формате
-
-### Пример шаблонов:
-- Отчет о геодезических работах (с автоматическим заполнением полей)
-- Картограмма участка (с автоматическим расчетом площадей и номенклатур)
