--- docs/docs/scale.md
@@ -1,27 +1,0 @@
-# Масштабы
-
-Раздел для расчёта и подбора масштабов. Позволяет вычислять масштаб по расстояниям и площадям, выбирать стандартные значения, получать точность.
-
-## Основные функции
-- Расчёт масштаба по расстояниям (реальное и на карте)
-- Расчёт масштаба по площадям
-- Ручной ввод масштаба (например, '1:5000')
-- Расчёт расстояний по масштабу
-- Выбор стандартных масштабов
-- Получение точности для выбранного масштаба
-
-## Используемые сервисы и утилиты
-- `ScaleService` — бизнес-логика расчётов
-- `ScaleUtils` — вспомогательные функции и стандартные значения
-
-## Пример использования
-```python
-from src.services.scale_service import ScaleService
-result = ScaleService.calculate_scale_from_distance(1000, 20)
-print(result.scale_text)  # '1:50'
-```
-
-## Сценарии применения
-- Автоматизация расчёта масштаба для карт и чертежей
-- Проверка соответствия масштаба стандартам
-- Получение точности для отчётов и документации
