--- docs/docs/service_coordinate.md
@@ -1,8 +1,0 @@
-# Сервис: Координаты
-
-`CoordinateService` — сервис для работы с координатами.
-
-## Основные функции
-- Парсинг координат из текста
-- Форматирование координат
-- Проверка формата и преобразование
