--- docs/docs/settings.md
@@ -1,19 +1,0 @@
-# Настройки
-
-Раздел для управления настройками приложения GeoOffice. Позволяет настраивать поведение, внешний вид и параметры хранения данных.
-
-## Основные функции
-- Включение/отключение автосохранения данных
-- Переключение темы (светлая/тёмная)
-- Экспорт и импорт пользовательских данных и настроек
-- Сброс настроек к значениям по умолчанию
-- Просмотр информации о версии и дате сборки
-
-## Используемые сервисы и утилиты
-- `FileUtils` — экспорт и импорт данных
-- `SettingsService` — управление настройками приложения
-
-## Пример сценария
-1. Включить автосохранение для предотвращения потери данных
-2. Переключить тему интерфейса под предпочтения пользователя
-3. Экспортировать настройки для переноса на другой компьютер
