--- docs/docs/taxation.md
@@ -1,29 +1,0 @@
-# Таксация
-
-Раздел для выполнения таксационных расчетов — автоматизация расчетов по участкам, деревьям, ведомостям.
-
-## Основные функции
-- Расчет отходов по данным участка и деревьев
-- Проверка и валидация ведомости
-- Ввод и редактирование данных по участку и деревьям
-- Автоматизация расчетов и формирование отчетов
-
-## Используемые сервисы и утилиты
-- `TaxationService` — бизнес-логика расчетов
-- `FileUtils` — загрузка и сохранение ведомостей
-
-## Пример сценария
-1. Ввести или импортировать данные по участку и деревьям
-2. Выполнить расчет отходов и проверить ведомость
-3. Экспортировать результаты для отчетности
-
-## Пример структуры данных
-```json
-{
-  "участок": "№1",
-  "деревья": [
-    {"порода": "сосна", "диаметр": 30, "высота": 20},
-    {"порода": "ель", "диаметр": 25, "высота": 18}
-  ]
-}
-```
