+++ src/components/banners.py
@@ -4,7 +4,7 @@ import flet as ft
 
 from src.utils.logger_config import get_logger
 
-logger = get_logger("components.categorized_menu")
+logger = get_logger("components.banners")
 
 
 class BannerDiffProjects:
